Refactor. split URLImageTile and XFileImageTile into separate file.
