+++ lib/presentation/components/layout_images.dart
@@ -1,7 +1,5 @@
-import 'package:cached_network_image/cached_network_image.dart';
+import 'package:farmers_journal/presentation/components/image_tile.dart';
 import 'package:flutter/material.dart';
-import 'package:image_picker/image_picker.dart';
-import 'dart:io';
 import 'package:farmers_journal/presentation/pages/page_journal/image_type.dart';
 import 'package:farmers_journal/presentation/components/layout_images_detail_screen.dart';
 
@@ -487,7 +485,7 @@ class CustomImageWidgetLayout extends StatelessWidget {
           isEditMode: isEditMode,
           borderRadius: borderRadius,
         ),
-      XFileImage(:final value) => _XFileImageTile(
+      XFileImage(:final value) => XFileImageTile(
           id: index,
           image: value,
           onDelete: () => onDelete?.call(index),
@@ -538,133 +536,5 @@ class CustomImageWidgetLayout extends StatelessWidget {
       case XFileImage():
         return imageTile;
     }
-  }
-}
-
-/// Displays network images with caching support via CachedNetworkImage. Shows delete button in edit mode.
-class URLImageTile extends StatelessWidget {
-  const URLImageTile({
-    super.key,
-    required this.url,
-    required this.onDelete,
-    required this.isEditMode,
-    required this.maxWidth,
-    required this.maxHeight,
-    required this.minWidth,
-    required this.minHeight,
-    this.borderRadius,
-  });
-
-  final String url;
-  final void Function()? onDelete;
-  final bool isEditMode;
-  final double maxWidth;
-  final double maxHeight;
-  final double minWidth;
-  final double minHeight;
-  final BorderRadius? borderRadius;
-
-  @override
-  Widget build(BuildContext context) {
-    return Stack(children: [
-      ConstrainedBox(
-        constraints: BoxConstraints(
-          minHeight: minHeight,
-          minWidth: minWidth,
-          maxWidth: maxWidth,
-          maxHeight: maxHeight,
-        ),
-        child: ClipRRect(
-          borderRadius: borderRadius ??
-              const BorderRadius.only(
-                bottomLeft: Radius.zero,
-                bottomRight: Radius.zero,
-                topLeft: Radius.circular(10),
-                topRight: Radius.circular(10),
-              ),
-          child: CachedNetworkImage(
-            imageUrl: url,
-            fit: BoxFit.cover,
-            errorWidget: (context, url, error) => const Icon(
-              Icons.broken_image,
-              size: 50,
-            ),
-          ),
-        ),
-      ),
-      isEditMode
-          ? Positioned(
-              right: 1,
-              child: IconButton(
-                alignment: Alignment.topRight,
-                onPressed: onDelete,
-                padding: EdgeInsets.zero,
-                iconSize: 20,
-                icon: const Icon(
-                  Icons.cancel_rounded,
-                  color: Colors.white,
-                ),
-              ),
-            )
-          : const SizedBox.shrink(),
-    ]);
-  }
-}
-
-/// Displays local images selected from device storage. Shows delete button in edit mode.
-class _XFileImageTile extends StatelessWidget {
-  const _XFileImageTile({
-    required this.id,
-    required this.image,
-    required this.isEditMode,
-    required this.onDelete,
-    this.width,
-    this.height,
-    this.borderRadius,
-  });
-  final int id;
-  final XFile image;
-  final bool isEditMode;
-  final void Function()? onDelete;
-  final double? width;
-  final double? height;
-  final BorderRadius? borderRadius;
-  @override
-  Widget build(BuildContext context) {
-    final imageRect = ClipRRect(
-      borderRadius: borderRadius ?? BorderRadius.circular(10),
-      child: SizedBox(
-        width: width,
-        height: height,
-        child: Image.file(
-          File(image.path),
-          width: width,
-          height: height,
-          fit: BoxFit.cover,
-          errorBuilder: (context, error, stackTrace) {
-            return const Center(child: Icon(Icons.broken_image, size: 50));
-          },
-        ),
-      ),
-    );
-    final removeImageButton = Positioned(
-      right: 1,
-      child: IconButton(
-        alignment: Alignment.topRight,
-        onPressed: onDelete,
-        padding: EdgeInsets.zero,
-        iconSize: 20,
-        icon: const Icon(
-          Icons.cancel_rounded,
-          color: Colors.white,
-        ),
-      ),
-    );
-    return isEditMode
-        ? Stack(children: [
-            imageRect,
-            removeImageButton,
-          ])
-        : imageRect;
   }
 }
